chore(Refractor): Moved composites into `composites` subdirectory.
