+++ src/fold/composites/__init__.py
@@ -1,0 +1,7 @@
+from .base import Composite
+from .ensemble import Ensemble, PerColumnEnsemble
+from .hybrid import Hybrid
+from .metalabeling import MetaLabeling
+from .sample import Sample
+from .select import SelectBest
+from .target import TransformTarget
