+++ src/fold/composites/metalabeling.py
@@ -7,12 +7,8 @@ import pandas as pd
 from fold.transformations.common import get_concatenated_names
 from fold.utils.checks import get_prediction_column
 
-from ..transformations.base import (
-    BlocksOrWrappable,
-    Composite,
-    T,
-    TransformationsAlwaysList,
-)
+from ..transformations.base import BlocksOrWrappable, TransformationsAlwaysList
+from .base import Composite, T
 
 
 class MetaLabeling(Composite):
