+++ tests/test_sampling.py
@@ -1,8 +1,8 @@
 from imblearn.under_sampling import RandomUnderSampler
 
+from fold.composites.sample import Sample
 from fold.loop import train
 from fold.splitters import ExpandingWindowSplitter
-from fold.transformations.sample import Sample
 from fold.transformations.test import Test
 from fold.utils.tests import generate_zeros_and_ones_skewed
