+++ tests/test_transform_target.py
@@ -1,10 +1,10 @@
 import numpy as np
 
+from fold.composites.target import TransformTarget
 from fold.loop import backtest, train
 from fold.models.dummy import DummyRegressor
 from fold.splitters import ExpandingWindowSplitter
 from fold.transformations.difference import Difference
-from fold.transformations.target import TransformTarget
 from fold.transformations.test import Test
 from fold.utils.tests import generate_monotonous_data, generate_zeros_and_ones_skewed
